move applyScales
