+++ src/mark.js
@@ -312,23 +312,6 @@ export function numberChannel(source) {
   };
 }
 
-// TODO use Float64Array.from for position and radius scales?
-export function applyScales(channels = [], scales) {
-  const values = Object.create(null);
-  for (let [name, {value, scale}] of channels) {
-    if (name !== undefined) {
-      if (scale !== undefined) {
-        scale = scales[scale];
-        if (scale !== undefined) {
-          value = Array.from(value, scale);
-        }
-      }
-      values[name] = value;
-    }
-  }
-  return values;
-}
-
 export function isOrdinal(values) {
   for (const value of values) {
     if (value == null) continue;
